chore: nostr hook
