+++ app/sign-in/page.tsx
@@ -11,6 +11,7 @@ import { useDispatch } from 'react-redux'
 
 import { Button } from '@/components/Button'
 import Layout from '@/components/Layout'
+import { relays } from '@/constants/nostr'
 import { routes } from '@/constants/routes'
 import { login } from '@/redux/features/user'
 
@@ -20,10 +21,7 @@ const SignIn = () => {
   const router = useRouter()
   const dispatch = useDispatch()
 
-  const ndk = useMemo(
-    () => new NDK({ explicitRelayUrls: ['wss://relay.primal.net'] }),
-    [],
-  )
+  const ndk = useMemo(() => new NDK({ explicitRelayUrls: relays }), [])
 
   const newLogin = useCallback(async () => {
     const signer = !nsec
