+++ constants/nostr.ts
@@ -1,0 +1,6 @@
+// TODO: maybe this will be a state in the future where the user can add and remove their own relays
+export const relays = ['wss://relay.primal.net']
+
+export enum APP_SPECIFIC_TAGS {
+  EMERGENCY_CONTACT = 'emergency-contact',
+}
